fstab: mount the partition with the boot files at /boot
Mount the partition /dev/boot-part on /boot.
This requires the help of udev/scripts/mount.sh. The script checks
the /proc/cmdline for the rootfs partition. If this is the second
partition of a block device and a first partition exists it creates
the symlink /dev/boot-part.

This reverts the previous mounting scheme which unconditionally mounted
the first eMMC partition at /boot, independent of the eMMC being the
actual device one booted from. commit 38705bd ("fstab: configure
the emmc's first partition to mount at /boot")

Related-to: ELB-3323
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>